% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_gt_clumppling.R
\name{autoplot.gt_clumppling}
\alias{autoplot.gt_clumppling}
\title{autoplot for clumppling objects}
\usage{
\method{autoplot}{gt_clumppling}(
  object,
  type = c("modes", "modes_within_k", "major_modes", "all_modes"),
  group = NULL,
  k = NULL,
  ...
)
}
\arguments{
\item{object}{a \code{\link{gt_clumppling}} object}

\item{type}{the type of plot, one of 'modes', 'modes_within_K', 'major_modes'
or 'all_modes'.}

\item{group}{a vector of membership to a-priori groups (e.g. populations).
Note that individuals from the same group need to be adjacent to each other}

\item{k}{the k value to be plotted if 'type' is 'modes_within_k'}

\item{...}{not used at the moment}
}
\value{
a plot
}
\description{
An autoplot method to generate quick visualisations for \code{\link{gt_clumppling}}
objects. Available types are:
\itemize{
\item 'modes': all aligned modes in structure plots over a multipartite graph,
where better alignment between the modes is indicated by the darker color of
the edges connecting their structure plots, and the cost of optimal alignment
is labelled on each edge.
\item 'modes_within_K': A set of figures, one for each number of clusters, with
all modes with the same number of clusters in structure plots in one figure.
\item 'major_modes': the major modes of each K aligned in a series of structure
plots.
\item 'all_modes': all aligned modes in a series of structure plots.
}
}
\details{
\code{autoplot} produces simple plots to quickly inspect an object. They are not
customisable; we recommend that you use \code{ggplot2} to produce publication
ready plots.

If you would like to generate an annotated autoplot, ensure that all
individuals from the same population are adjacent to one another in the
Q-matrix or gt_admix object supplied to gt_clumppling. Autoplot 'group'
argument requires that all individuals from the same group are adjacent.
}
