% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_docs.R
\docType{data}
\name{capeverde_pops}
\alias{capeverde_pops}
\title{Population membership for the Capeverde dataset}
\format{
A vector of length 399
}
\usage{
capeverde_pops
}
\description{
Membership to populations for a dataset of 399 individuals, including 44
Cape Verdean. This dataset is used as an example for clumppling.
}
\keyword{datasets}
