% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumppling.R
\name{clumppling}
\alias{clumppling}
\title{run clumppling}
\usage{
clumppling(
  input_path,
  input_format = "admixture",
  cd_param = 1,
  use_rep = 0,
  merge_cls = 0,
  cd_default = 1,
  custom_cmap = "",
  output_path = tempfile("clumpling")
)
}
\arguments{
\item{input_path}{the path where the Q files are stored, either a directory
or a zip archive}

\item{input_format}{a string defining the format of the input files, one of
'admixture' (default)}

\item{cd_param}{the cd_param
use_rep boolean on whether a representative repeat should be used as a consesus
for a mode. Defaults to FALSE, which leads to the computatoin of an average}

\item{use_rep}{boolean}

\item{merge_cls}{boolean,}

\item{cd_default}{boolean}

\item{custom_cmap}{a custom colour map (this should be moved
to the plotting functions!) a string that is either empty (default “”)
or contains a list of colors to be used for a customized colormap. If
the string is empty then the default colormap will be used.
The customized colormap, if provided, should be a list of colors in hex
code in a comma-delimited string. An example colormap with five
colors looks like "#FF0000,#FFFF00,#00EAFF #AA00FF,#FF7F00". If
the provided colormap does not have enough colors for all clusters,
colors will be cycled.}

\item{output_path}{(optional) the clumppling functions in python save
everything to file. By default, R stores the information in objects in the
environment, and sends those files to a temporary directory
that will be cleared at the end of a session. \code{output_path} allows to
change the location
of those files. This is only useful to those interested in recovering the
same files as created by the python clumppling module, or for debugging.}
}
\value{
a list of class \code{clumpling} TODO describe the elements of the list
}
\description{
This function runs the clumppling algorithm.
}
