% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastmixture.R
\name{fastmixture}
\alias{fastmixture}
\title{fastmixture algorithm for population genetics clustering}
\usage{
fastmixture(
  bfile,
  k,
  threads = 1,
  seed = 42,
  outprefix = "fastmixture",
  iter = 1000,
  tole = 0.5,
  batches = 32,
  supervised = NULL,
  check = 5,
  power = 11,
  chunk = 8192,
  als_iter = 1000,
  als_tole = 1e-04,
  no_freqs = TRUE,
  random_init = TRUE
)
}
\arguments{
\item{bfile}{is the name of the binary plink file (without the .bed extension)}

\item{k}{the number of ancestral components (clusters)}

\item{threads}{the number of threads to use (1)}

\item{seed}{the random seed (42)}

\item{outprefix}{the prefix of the output files (fastmixture)}

\item{iter}{the maximum number of iterations (1000)}

\item{tole}{the tolerance in log-likelihood units between iterations (0.5)}

\item{batches}{the number of maximum mini-batches (32)}

\item{supervised}{the name fo the file with the supervised labels (NULL)}

\item{check}{the number of iterations to check for convergence (5)}

\item{power}{number of power iterations in randomised SVD (11)}

\item{chunk}{the number of SPs in chunk operations (8192)}

\item{als_iter}{the maximum number of iterations in the ALS algorithm (1000)}

\item{als_tole}{the tolerance for the RMSE of P between iterations (1e-4)}

\item{no_freqs}{do not save P-matrix (TRUE)}

\item{random_init}{random initialisation of parameters (TRUE)}
}
\value{
NULL (for the moment, in the future a list)
}
\description{
This function implements the fastmixture algorithm for population genetics clustering
by calling the relevant python module.
}
