% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastmixture_k_list.R
\name{fastmixture_k_list}
\alias{fastmixture_k_list}
\title{Run fastmixture for multiple K values}
\usage{
fastmixture_k_list(
  gt_object,
  k_values,
  threads = 1,
  seed = 42,
  iter = 1000,
  tole = 0.5,
  batches = 32,
  supervised = NULL,
  check = 5,
  power = 11,
  chunk = 8192,
  als_iter = 1000,
  als_tole = 1e-04,
  no_freqs = TRUE,
  random_init = TRUE,
  save_q = FALSE,
  output_path = getwd()
)
}
\arguments{
\item{gt_object}{A gentibble object that will be analyzed by the fastmixture algorithm.}

\item{k_values}{A vector of integers specifying the different K values for which to run the \code{fastmixture}.}

\item{threads}{Number of threads to use for parallelization. Default is 1.}

\item{seed}{Random seed for reproducibility. Default is 42.}

\item{iter}{Number of iterations for the fastmixture algorithm. Default is 1000.}

\item{tole}{Convergence tolerance value for the fastmixture algorithm. Default is 0.5.}

\item{batches}{Number of batches for the fastmixture algorithm. Default is 32.}

\item{supervised}{Supervised parameter for fastmixture. Default is NULL.}

\item{check}{Number of checks to be performed during the process. Default is 5.}

\item{power}{Power parameter used for the fastmixture algorithm. Default is 11.}

\item{chunk}{Size of data chunk used in processing. Default is 8192.}

\item{als_iter}{Number of ALS (alternating least squares) iterations. Default is 1000.}

\item{als_tole}{Tolerance value for ALS convergence. Default is 1e-4.}

\item{no_freqs}{Logical parameter to indicate whether to use allele frequencies. Default is TRUE.}

\item{random_init}{Logical parameter to indicate whether to use random initialization. Default is TRUE.}

\item{save_q}{Logical parameter indicating whether to save each Q matrix as a file. Default is \code{FALSE}.}

\item{output_path}{The path where Q files should be saved if \code{save_q} is \code{TRUE}. Default is the current working directory.}
}
\value{
A list of \code{q_matrix} objects, one for each K value provided.
}
\description{
This function runs the \code{fastmixture} algorithm for a range of K values and returns
a list of \code{q_matrix} objects for each K.
}
\examples{
\dontrun{
k_values <- c(2, 3, 4)
result <- fastmixture_k_list(agt_object, k_values)
}
}
