% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_clumppling.R
\name{gt_clumppling}
\alias{gt_clumppling}
\title{run clumppling}
\usage{
gt_clumppling(
  input_path,
  output_path = tempfile("clump_out"),
  input_format = "admixture",
  use_rep = TRUE,
  merge = TRUE,
  cd_method = "louvain",
  use_best_pair = TRUE,
  extension = ".Q"
)
}
\arguments{
\item{input_path}{the path where the Q files are stored, either a directory
or a zip archive, or a \code{q_matrix_list} object}

\item{output_path}{(optional) the clumppling functions in python save
everything to file. By default, R stores the information in objects in the
environment, and sends those files to a temporary directory that will be
cleared at the end of a session. \code{output_path} allows to change the
location of those files. This is only useful to those interested in
recovering the same files as created by the python clumppling module, or
for debugging.}

\item{input_format}{a string defining the format of the input files, one of
'admixture' (default),'structure','fastStructure' or 'generalQ'}

\item{use_rep}{boolean, whether to use representative modes (alternative:
average), defaults to TRUE}

\item{merge}{boolean, whether to merge two clusters when aligning K+1 to K,
defaults to TRUE}

\item{cd_method}{the community detection method to use, one of 'louvain'
(default), 'leiden', 'infomap', 'markov_clustering', 'label_propagation',
'walktrap', 'custom'}

\item{use_best_pair}{boolean, whether to use best pair as anchor for across-K
alignment (alternative: major), defaults to TRUE}

\item{extension}{(optional) if loading from files rather than a
\code{q_matrix_list} object specify the extension e.g. ".Q" or ".indivq"}
}
\value{
a list of class \code{gt_clumppling} containing:
\itemize{
\item N: number of individuals
\item K_range: vector of K values analyzed
\item mode_replicates: a list of replicate indices for each mode
\item cost_acrossK: a named list of costs for each pairwise K alignment
\item aligned_modes: a list of data.frames, each data.frame is a Q-matrix
}
}
\description{
This function runs the clumppling algorithm.
}
\details{
If you would like to generate an annotated autoplot from your gt_clumppling
object, ensure that all individuals from the same population are adjacent to
one another in the Q-matrix or gt_admix object supplied to gt_clumppling.
Autoplot 'group' argument requires that all individuals from the same group
are adjacent.
}
