% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygenclust-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{augment}
\alias{autoplot}
\alias{theme_distruct}
\alias{scale_fill_distruct}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{tidy}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{tidypopgen}{\code{\link[tidypopgen]{scale_fill_distruct}}, \code{\link[tidypopgen]{theme_distruct}}}
}}

