% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_clumppling.R
\name{subset_gt_clumppling}
\alias{subset_gt_clumppling}
\title{Subset a \code{gt_clumppling} object}
\usage{
subset_gt_clumppling(x, k = NULL, indivs = NULL)
}
\arguments{
\item{x}{a gt_clumppling object}

\item{k}{a vector of k values to subset to}

\item{indivs}{a vector of individual indices to keep}
}
\value{
a gt_clumppling object subsetted to the individuals specified
}
\description{
This function subsets \code{gt_clumppling} objects to a set of individuals or a
set of values of K. This is intended to create plot insets, or to visualise a
subset of individuals during data analysis. To understand the modes within a
subset of individuals in your data, you should subset your \code{gt_admix} object
and re-run \code{gt_clumppling}.
}
