% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_clumppling.R
\name{tidy.gt_clumppling}
\alias{tidy.gt_clumppling}
\title{Tidy a \code{gt_clumppling} object}
\usage{
\method{tidy}{gt_clumppling}(
  x,
  matrix = c("modes", "major_modes", "Q_modes", "q_modes", "Q_major_modes",
    "q_major_modes"),
  ...
)
}
\arguments{
\item{x}{the \code{\link{gt_clumppling}} object}

\item{matrix}{a string defining the information to be extracted, one of:
"modes", "major_modes", "Q_modes", "Q_major_modes".}

\item{...}{Additional arguments. Not used. Needed to match generic signature
only.}
}
\value{
a \link[tibble:tibble]{tibble::tibble} of the information of interest
}
\description{
A \code{tidy} method to extract information from a \code{\link{gt_clumppling}} object, and
return it as a tibble. It can extract:
\itemize{
\item 'modes': all the modes detected by \code{\link[=gt_clumppling]{gt_clumppling()}}. The models have label
'KxMy', where 'x' and 'y' represent the K value and the mode rank.
\item 'major_modes': modes of rank 1 for each K.
\item 'Q_modes', 'q_modes': a list of q matrices, one per mode, each tidied
into a tibble
\item 'Q_major_modes', 'q_major_modes': the same output as 'Q_modes' but
subsetted to only the
major modes.
}
}
