% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastmixture.R
\name{fastmixture}
\alias{fastmixture}
\title{fastmixture algorithm for population genetics clustering}
\usage{
fastmixture(
  data,
  k,
  threads = 1,
  seed = 42,
  outprefix = "fastmixture",
  iter = 1000,
  tole = 0.5,
  batches = 32,
  supervised = NULL,
  check = 5,
  power = 11,
  chunk = 8192,
  als_iter = 1000,
  als_tole = 1e-04,
  no_freqs = TRUE,
  random_init = TRUE
)
}
\arguments{
\item{data}{either a \code{\link[tidypopgen:gen_tibble]{tidypopgen::gen_tibble}}, or the name of the binary plink file (without the .bed extension)}

\item{k}{the number of ancestral components (clusters)}

\item{threads}{the number of threads to use (1)}

\item{seed}{the random seed (42)}

\item{outprefix}{the prefix of the output files (fastmixture)}

\item{iter}{the maximum number of iterations (1000)}

\item{tole}{the tolerance in log-likelihood units between iterations (0.5)}

\item{batches}{the number of maximum mini-batches (32)}

\item{supervised}{the name fo the file with the supervised labels (NULL)}

\item{check}{the number of iterations to check for convergence (5)}

\item{power}{number of power iterations in randomised SVD (11)}

\item{chunk}{the number of SPs in chunk operations (8192)}

\item{als_iter}{the maximum number of iterations in the ALS algorithm (1000)}

\item{als_tole}{the tolerance for the RMSE of P between iterations (1e-4)}

\item{no_freqs}{do not save P-matrix (TRUE)}

\item{random_init}{random initialisation of parameters (TRUE)}
}
\value{
either the q matrix (if no_freqs=TRUE; formatted as a \code{tidypopgen::q_matrix}) or a list of the Q and P matrices
}
\description{
This function implements the fastmixture algorithm for population genetics clustering
by calling the python module. If you use this function, make sure that you cite
the relevant paper by Santander, Refoyo-Martínez, and Meisner (2024).
}
\details{
This function returns a q_matrix that can be plotted with \code{autoplot}, and tidied with \code{tidy}
methods from the \code{tidypopgen} package.
}
\references{
C. G. Santander, A. Refoyo Martinez, J. Meisner (2024) Faster model-based estimation of ancestry proportions. bioRxiv 2024.07.08.602454; doi: https://doi.org/10.1101/2024.07.08.602454
}
