% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_clumppling.R
\name{gt_clumppling}
\alias{gt_clumppling}
\title{run clumppling}
\usage{
gt_clumppling(
  input_path,
  input_format = "admixture",
  cd_param = 1,
  use_rep = 0,
  merge_cls = 0,
  cd_default = 1,
  output_path = tempfile("clump_out")
)
}
\arguments{
\item{input_path}{the path where the Q files are stored, either a directory
or a zip archive, or a \code{q_matrix_list} object}

\item{input_format}{a string defining the format of the input files, one of
'admixture' (default)}

\item{cd_param}{the cd_param
use_rep boolean on whether a representative repeat should be used as a consesus
for a mode. Defaults to FALSE, which leads to the computatoin of an average}

\item{use_rep}{boolean}

\item{merge_cls}{boolean,}

\item{cd_default}{boolean}

\item{output_path}{(optional) the clumppling functions in python save
everything to file. By default, R stores the information in objects in the
environment, and sends those files to a temporary directory
that will be cleared at the end of a session. \code{output_path} allows to
change the location
of those files. This is only useful to those interested in recovering the
same files as created by the python clumppling module, or for debugging.}
}
\value{
a list of class \code{clumpling} TODO describe the elements of the list
}
\description{
This function runs the clumppling algorithm.
}
\details{
If you would like to generate an annotated autoplot from your gt_clumppling object,
ensure that all individuals from the same population are adjacent to one another
in the Q-matrix or gt_admix object supplied to gt_clumppling. Autoplot 'group'
argument requires that all individuals from the same group are adjacent.
}
