% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgc_tools_install.R
\name{tgc_tools_install}
\alias{tgc_tools_install}
\title{Install tools for \code{tidygenclust}}
\usage{
tgc_tools_install(envname = "ctidygenclust", reset = FALSE)
}
\arguments{
\item{envname}{the name of the virtual environment to use (defaults to 'ctidygenclust')}

\item{reset}{a boolean used to reset the virtual environment. Only set
to TRUE if you have a broken virtual environment that you want to reset.}
}
\description{
\code{tidygenclust} relies on the python packages \code{fastmixture} and \code{clumppling} for a number of
functionalities. We use \code{reticulate} to install them in a conda
environment. We also install the other tools in the same environment to
make sure that we have everything in the same place.  By default, we use a
conda environment called 'ctidygenclust'. Unless
you know what you are doing, we recommend that you use this default. For
more details, see \href{https://rstudio.github.io/reticulate/articles/python_dependencies.html}{here}.
}
