% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygenclust-package.R
\docType{package}
\name{tidygenclust-package}
\alias{tidygenclust}
\alias{tidygenclust-package}
\title{tidygenclust: Clustering for Population Genetics in R}
\description{
A tidy interface to clustering in population genetics. This package provides a set of functions to perform clustering on genetic data, and to visualize the results, both for single runs and for multiple repeats of the same analysis. It is built on top of the 'tidypopgen' package, but can also use PLINK input files.
}
\author{
\strong{Maintainer}: Andrea Manica \email{am315@cam.ac.uk} (\href{https://orcid.org/0000-0003-1895-450X}{ORCID})

}
\keyword{internal}
